refactor(tasks): export getIosProjectPath
